+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+## Version 1.7.1
+- Updated policy cmdlets to use new api version 2019-06-01 that has new EnforcementMode property in policy assignment.
+- Updated create policy definition help example
+
 ## Version 1.7.0
 * Fix bug where New-AzRoleAssignment could not be called without parameter Scope.